Refactor Symbol Analysis page
