+++ src/app/symbol/[ticker]/page.tsx
@@ -2,13 +2,10 @@
 
 import { useParams, useRouter } from "next/navigation";
 import Link from "next/link";
-import { useState, useEffect, useCallback, useMemo } from "react";
+import { useState, useEffect, useMemo } from "react";
 import { Option } from "effect";
 import { useAuth } from "@/contexts/AuthContext";
-import type { InsiderTransactionsDBRow, ValuationsDBRow, GradesHistoricalDBRow, AnalystPriceTargetsDBRow } from "@/lib/supabase/realtime-service";
-import type { Database } from "@/lib/supabase/database.types";
 
-type FinancialStatementDBRow = Database["public"]["Tables"]["financial_statements"]["Row"];
 import { Card, CardContent, CardHeader, CardTitle, CardDescription } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
@@ -23,15 +20,12 @@ import Image from "next/image";
 import { useAddCardToWorkspace } from "@/hooks/useAddCardToWorkspace";
 import { useWorkspaceCards, removeSymbolFromWorkspace } from "@/hooks/useWorkspaceCards";
 import type { CardType } from "@/components/game/cards/base-card/base-card.types";
-import { useStockData, type ProfileDBRow } from "@/hooks/useStockData";
-import {
-  type MarketRiskPremiumDBRow,
-  type TreasuryRateDBRow,
-  type MarketRiskPremiumPayload,
-} from "@/lib/supabase/realtime-service";
+import { useSymbolAnalysisData } from "@/hooks/useSymbolAnalysisData";
+import { MetricRow } from "@/components/symbol/MetricRow";
+import { ScorecardItem } from "@/components/symbol/ScorecardItem";
+
 import { formatFinancialValue } from "@/lib/formatters";
 import { useExchangeRate } from "@/hooks/useExchangeRate";
-import type { RatiosTtmDBRow } from "@/lib/supabase/realtime-service";
 import {
   calculateROIC,
   calculateFCFYield,
@@ -39,6 +33,12 @@ import {
   calculateAltmanZScore,
   calculateInterestCoverage,
 } from "@/lib/financial-calculations";
+import {
+  calculateValuationStatus,
+  calculateQualityStatus,
+  calculateSafetyStatus,
+  calculateContrarianIndicatorsStatus,
+} from "@/lib/status-calculations";
 import {
   Line,
   ResponsiveContainer,
@@ -97,459 +97,6 @@ interface ContrarianIndicators {
   priceTarget: Option.Option<number>;
 }
 
-type HealthStatus = "Undervalued" | "Fair" | "Overvalued" | "Unknown";
-type QualityStatus = "Excellent" | "Good" | "Average" | "Poor" | "Unknown";
-type SafetyStatus = "Safe" | "Moderate" | "Risky" | "Unknown";
-
-// ============================================================================
-// CALCULATION HELPERS
-// ============================================================================
-
-function calculateValuationStatus(
-  price: Option.Option<number>,
-  dcf: Option.Option<number>,
-  peRatio: Option.Option<number>,
-  pegRatio: Option.Option<number>
-): { status: HealthStatus; color: string; borderColor: string } {
-  // If we don't have enough data, return unknown
-  if (Option.isNone(price) || Option.isNone(dcf)) {
-    return { status: "Unknown", color: "text-muted-foreground", borderColor: "border-l-border" };
-  }
-
-  const priceVal = price.value;
-  const dcfVal = dcf.value;
-
-  // Edge case: If DCF is negative or zero, treat as overvalued (distressed company)
-  if (dcfVal <= 0 || !isFinite(dcfVal)) {
-    return { status: "Overvalued", color: "text-red-600", borderColor: "border-l-red-500" };
-  }
-
-  const discount = ((dcfVal - priceVal) / dcfVal) * 100;
-
-  // Score-based approach: Consider DCF discount, P/E, and PEG
-  let score = 0;
-  let signals = 0;
-
-  // Signal 1: DCF Discount (weight: 40%)
-  // > 20% discount = undervalued, < -20% = overvalued
-  if (discount > 20) {
-    score += 40;
-    signals++;
-  } else if (discount < -20) {
-    score -= 40;
-    signals++;
-  } else if (discount > 0) {
-    score += 20; // Slightly undervalued
-    signals++;
-  } else {
-    score -= 20; // Slightly overvalued
-    signals++;
-  }
-
-  // Signal 2: P/E Ratio (weight: 30%)
-  // Thresholds are market-average based (15, 20, 25, 30)
-  // - < 15: Very Cheap (below market average)
-  // - 15-20: Reasonable (market average range)
-  // - 20-25: Fair Value (slightly above average)
-  // - 25-30: Expensive (high growth expectations)
-  // - > 30: Very Expensive (very high growth or overvaluation)
-  // Note: Industry-adjusted thresholds would improve accuracy for:
-  //   - Technology companies (typically 20-40 P/E, may be fairly valued at 25-30)
-  //   - Financial companies (typically 10-15 P/E, may be fairly valued at 12-15)
-  // Current approach works for general-purpose analysis across most industries
-  // Edge case: Negative P/E (loss-making companies) = very expensive (no earnings yield)
-  if (Option.isSome(peRatio)) {
-    const pe = peRatio.value;
-    // Handle negative P/E (loss-making companies) - treat as very expensive
-    if (pe <= 0 || !isFinite(pe)) {
-      score -= 30;
-      signals++;
-    } else if (pe < 15) {
-      score += 30;
-      signals++;
-    } else if (pe < 20) {
-      score += 15;
-      signals++;
-    } else if (pe > 30) {
-      score -= 30;
-      signals++;
-    } else if (pe > 25) {
-      score -= 15;
-      signals++;
-    } else {
-      signals++; // Neutral zone (20 <= pe <= 25)
-    }
-  }
-
-  // Signal 3: PEG Ratio (weight: 30%)
-  // PEG < 1.0 = undervalued, 1.0-2.0 = fair, > 2.0 = overvalued
-  // Edge case: Negative PEG (negative growth or earnings) = very overvalued
-  if (Option.isSome(pegRatio)) {
-    const peg = pegRatio.value;
-    // Handle negative PEG (negative growth or earnings) - treat as very overvalued
-    if (peg <= 0 || !isFinite(peg)) {
-      score -= 30;
-      signals++;
-    } else if (peg < 1.0) {
-      score += 30;
-      signals++;
-    } else if (peg < 1.5) {
-      score += 15;
-      signals++;
-    } else if (peg > 2.5) {
-      score -= 30;
-      signals++;
-    } else if (peg > 2.0) {
-      score -= 15;
-      signals++;
-    } else {
-      signals++; // Neutral zone (1.5 <= peg <= 2.0)
-    }
-  }
-
-  // Determine status based on composite score
-  // If we have at least 2 signals, use the score
-  // Otherwise, fall back to DCF-only logic
-  if (signals >= 2) {
-    if (score >= 30) {
-      return { status: "Undervalued", color: "text-green-600", borderColor: "border-l-green-500" };
-    } else if (score <= -30) {
-      return { status: "Overvalued", color: "text-red-600", borderColor: "border-l-red-500" };
-    } else {
-      return { status: "Fair", color: "text-yellow-600", borderColor: "border-l-yellow-500" };
-    }
-  } else {
-    // Fallback to DCF-only logic if we don't have enough signals
-    if (discount > 20) {
-      return { status: "Undervalued", color: "text-green-600", borderColor: "border-l-green-500" };
-    } else if (discount < -20) {
-      return { status: "Overvalued", color: "text-red-600", borderColor: "border-l-red-500" };
-    } else {
-      return { status: "Fair", color: "text-yellow-600", borderColor: "border-l-yellow-500" };
-    }
-  }
-}
-
-function calculateQualityStatus(
-  roic: Option.Option<number>,
-  wacc: Option.Option<number>,
-  grossMargin: Option.Option<number>,
-  fcfYield: Option.Option<number>,
-  roicHistory: { date: string; dateLabel: string; roic: number; wacc: number }[]
-): { status: QualityStatus; color: string; borderColor: string } {
-  // If we don't have ROIC, we can't assess quality
-  if (Option.isNone(roic)) {
-    return { status: "Unknown", color: "text-muted-foreground", borderColor: "border-l-border" };
-  }
-
-  const roicVal = roic.value;
-  let score = 0;
-  let signals = 0;
-
-  // Signal 1: ROIC vs WACC Spread (40% weight) - Most important indicator
-  // ROIC > WACC means the company is creating value
-  if (Option.isSome(wacc)) {
-    const waccVal = wacc.value;
-    const spread = roicVal - waccVal; // Percentage points difference
-
-    if (spread > 0.10) {
-      score += 40; // Exceptional value creation (>10% spread)
-      signals++;
-    } else if (spread > 0.05) {
-      score += 30; // Strong value creation (5-10% spread)
-      signals++;
-    } else if (spread > 0) {
-      score += 15; // Creating value (0-5% spread)
-      signals++;
-    } else if (spread > -0.05) {
-      score -= 15; // Marginally destroying value (-5% to 0% spread)
-      signals++;
-    } else {
-      score -= 40; // Significantly destroying value (<-5% spread)
-      signals++;
-    }
-  } else {
-    // If no WACC, use absolute ROIC thresholds
-    if (roicVal > 0.20) {
-      score += 30; // >20% ROIC is exceptional
-      signals++;
-    } else if (roicVal > 0.15) {
-      score += 20; // >15% ROIC is excellent
-      signals++;
-    } else if (roicVal > 0.10) {
-      score += 10; // >10% ROIC is good
-      signals++;
-    } else if (roicVal > 0.05) {
-      score += 0; // 5-10% ROIC is average
-      signals++;
-    } else if (roicVal > 0) {
-      score -= 20; // 0-5% ROIC is poor
-      signals++;
-    } else {
-      score -= 40; // Negative ROIC is destroying value
-      signals++;
-    }
-  }
-
-  // Signal 2: ROIC Absolute Level (25% weight)
-  if (roicVal > 0.20) {
-    score += 25; // Exceptional (>20%)
-    signals++;
-  } else if (roicVal > 0.15) {
-    score += 20; // Excellent (15-20%)
-    signals++;
-  } else if (roicVal > 0.10) {
-    score += 10; // Good (10-15%)
-    signals++;
-  } else if (roicVal > 0.05) {
-    score += 0; // Average (5-10%)
-    signals++;
-  } else if (roicVal > 0) {
-    score -= 15; // Poor (0-5%)
-    signals++;
-  } else {
-    score -= 30; // Negative (destroying value)
-    signals++;
-  }
-
-  // Signal 3: Gross Margin (20% weight) - Pricing power indicator
-  if (Option.isSome(grossMargin)) {
-    const margin = grossMargin.value;
-    if (margin > 0.60) {
-      score += 20; // Exceptional pricing power (>60%)
-      signals++;
-    } else if (margin > 0.40) {
-      score += 15; // Strong pricing power (40-60%)
-      signals++;
-    } else if (margin > 0.30) {
-      score += 5; // Moderate pricing power (30-40%)
-      signals++;
-    } else if (margin > 0.20) {
-      score -= 5; // Weak pricing power (20-30%)
-      signals++;
-    } else {
-      score -= 15; // Very weak pricing power (<20%)
-      signals++;
-    }
-  }
-
-  // Signal 4: FCF Yield (15% weight) - Cash generation
-  if (Option.isSome(fcfYield)) {
-    const fcfYieldValue = fcfYield.value;
-    if (fcfYieldValue > 0.10) {
-      score += 15; // Exceptional cash generation (>10%)
-      signals++;
-    } else if (fcfYieldValue > 0.05) {
-      score += 10; // Strong cash generation (5-10%)
-      signals++;
-    } else if (fcfYieldValue > 0.03) {
-      score += 5; // Moderate cash generation (3-5%)
-      signals++;
-    } else if (fcfYieldValue > 0) {
-      score -= 5; // Weak cash generation (0-3%)
-      signals++;
-    } else {
-      score -= 15; // Negative cash flow
-      signals++;
-    }
-  }
-
-  // Signal 5: ROIC Trend (bonus/penalty) - Is quality improving or declining?
-  if (roicHistory.length >= 2) {
-    const recent = roicHistory.slice(0, 3); // Last 3 data points
-    const oldest = recent[recent.length - 1];
-    const newest = recent[0];
-    const trend = newest.roic - oldest.roic; // Change in ROIC (already in percentage form)
-
-    if (trend > 0.05) {
-      score += 10; // Improving significantly (>5% points)
-    } else if (trend > 0.02) {
-      score += 5; // Improving moderately (2-5% points)
-    } else if (trend < -0.05) {
-      score -= 10; // Declining significantly (<-5% points)
-    } else if (trend < -0.02) {
-      score -= 5; // Declining moderately (-2 to -5% points)
-    }
-  }
-
-  // Determine status based on composite score
-  // Require at least 2 signals for a valid assessment
-  if (signals < 2) {
-    return { status: "Unknown", color: "text-muted-foreground", borderColor: "border-l-border" };
-  }
-
-  if (score >= 50) {
-    return { status: "Excellent", color: "text-green-600", borderColor: "border-l-green-500" };
-  } else if (score >= 20) {
-    return { status: "Good", color: "text-green-600", borderColor: "border-l-green-500" };
-  } else if (score >= -10) {
-    return { status: "Average", color: "text-yellow-600", borderColor: "border-l-yellow-500" };
-  } else {
-    return { status: "Poor", color: "text-red-600", borderColor: "border-l-red-500" };
-  }
-}
-
-function calculateSafetyStatus(
-  netDebtToEbitda: Option.Option<number>,
-  altmanZScore: Option.Option<number>,
-  interestCoverage: Option.Option<number>
-): { status: SafetyStatus; color: string; borderColor: string } {
-  // If we don't have at least one metric, we can't assess safety
-  if (Option.isNone(netDebtToEbitda) && Option.isNone(altmanZScore) && Option.isNone(interestCoverage)) {
-    return { status: "Unknown", color: "text-muted-foreground", borderColor: "border-l-border" };
-  }
-
-  let score = 0;
-  let signals = 0;
-
-  // Signal 1: Net Debt to EBITDA (40% weight) - Most important debt metric
-  // Lower is better - indicates ability to pay down debt
-  if (Option.isSome(netDebtToEbitda)) {
-    const ratio = netDebtToEbitda.value;
-    if (ratio < 1.0) {
-      score += 40; // Exceptional (< 1.0x) - Very low debt burden
-      signals++;
-    } else if (ratio < 2.0) {
-      score += 30; // Excellent (1.0-2.0x) - Low debt burden
-      signals++;
-    } else if (ratio < 3.0) {
-      score += 20; // Good (2.0-3.0x) - Safe level
-      signals++;
-    } else if (ratio < 5.0) {
-      score -= 10; // Moderate (3.0-5.0x) - Moderate risk
-      signals++;
-    } else if (ratio < 7.0) {
-      score -= 30; // Risky (5.0-7.0x) - High debt burden
-      signals++;
-    } else {
-      score -= 40; // Very Risky (> 7.0x) - Very high debt burden
-      signals++;
-    }
-  }
-
-  // Signal 2: Altman Z-Score (35% weight) - Bankruptcy risk indicator
-  // Higher is better - indicates financial stability
-  if (Option.isSome(altmanZScore)) {
-    const zScore = altmanZScore.value;
-    if (zScore > 3.0) {
-      score += 35; // Safe Zone (> 3.0) - Low bankruptcy risk
-      signals++;
-    } else if (zScore > 2.7) {
-      score += 20; // Good (2.7-3.0) - Low to moderate risk
-      signals++;
-    } else if (zScore > 1.81) {
-      score -= 10; // Grey Zone (1.81-2.7) - Moderate bankruptcy risk
-      signals++;
-    } else if (zScore > 1.0) {
-      score -= 30; // Distress Zone (1.0-1.81) - High bankruptcy risk
-      signals++;
-    } else {
-      score -= 40; // Critical (< 1.0) - Very high bankruptcy risk
-      signals++;
-    }
-  }
-
-  // Signal 3: Interest Coverage (25% weight) - Ability to service debt
-  // Higher is better - indicates ability to pay interest obligations
-  if (Option.isSome(interestCoverage)) {
-    const coverage = interestCoverage.value;
-    // Special case: 999 indicates perfect/infinite coverage (no interest expense)
-    // This is returned by calculateInterestCoverage when interestExpense <= 0
-    // Both 999 and >10x give +25 points, but 999 is explicitly handled first for clarity
-    if (coverage >= 999) {
-      score += 25; // Perfect coverage (no interest expense) - Exceptional
-      signals++;
-    } else if (coverage > 10.0) {
-      score += 25; // Exceptional (> 10x) - Very safe
-      signals++;
-    } else if (coverage > 5.0) {
-      score += 20; // Excellent (5-10x) - Safe
-      signals++;
-    } else if (coverage > 3.0) {
-      score += 10; // Good (3-5x) - Adequate
-      signals++;
-    } else if (coverage > 1.5) {
-      score -= 15; // Moderate (1.5-3x) - Tight coverage
-      signals++;
-    } else if (coverage > 0) {
-      score -= 30; // Risky (0-1.5x) - May struggle to pay interest
-      signals++;
-    } else {
-      score -= 40; // Critical (< 0) - Cannot cover interest
-      signals++;
-    }
-  }
-
-  // Determine status based on composite score
-  // Require at least 2 signals for a valid assessment
-  if (signals < 2) {
-    return { status: "Unknown", color: "text-muted-foreground", borderColor: "border-l-border" };
-  }
-
-  if (score >= 50) {
-    return { status: "Safe", color: "text-green-600", borderColor: "border-l-green-500" };
-  } else if (score >= 10) {
-    return { status: "Moderate", color: "text-yellow-600", borderColor: "border-l-yellow-500" };
-  } else {
-    return { status: "Risky", color: "text-red-600", borderColor: "border-l-red-500" };
-  }
-}
-
-function calculateContrarianIndicatorsStatus(
-  analystConsensus: Option.Option<string>,
-  priceTarget: Option.Option<number>,
-  currentPrice: Option.Option<number>
-): { status: string; color: string; borderColor: string } {
-  // If we don't have enough data, return neutral/unknown
-  if (Option.isNone(analystConsensus) && Option.isNone(priceTarget)) {
-    return { status: "Unknown", color: "text-muted-foreground", borderColor: "border-l-border" };
-  }
-
-  let bullishSignals = 0;
-  let bearishSignals = 0;
-
-  // Check analyst consensus
-  if (Option.isSome(analystConsensus)) {
-    const consensus = analystConsensus.value;
-    if (consensus === "Strong Buy" || consensus === "Buy") {
-      bullishSignals += consensus === "Strong Buy" ? 2 : 1;
-    } else if (consensus === "Sell" || consensus === "Strong Sell") {
-      bearishSignals += consensus === "Strong Sell" ? 2 : 1;
-    }
-  }
-
-  // Check price target vs current price
-  if (Option.isSome(priceTarget) && Option.isSome(currentPrice)) {
-    const target = priceTarget.value;
-    const current = currentPrice.value;
-    const upside = ((target - current) / current) * 100;
-
-    if (upside > 10) {
-      bullishSignals += 2; // Strong bullish signal
-    } else if (upside > 0) {
-      bullishSignals += 1; // Moderate bullish signal
-    } else if (upside < -10) {
-      bearishSignals += 2; // Strong bearish signal
-    } else if (upside < 0) {
-      bearishSignals += 1; // Moderate bearish signal
-    }
-  }
-
-  // Determine overall status
-  if (bullishSignals > bearishSignals && bullishSignals >= 2) {
-    return { status: "Bullish", color: "text-green-600", borderColor: "border-l-green-500" };
-  } else if (bearishSignals > bullishSignals && bearishSignals >= 2) {
-    return { status: "Bearish", color: "text-red-600", borderColor: "border-l-red-500" };
-  } else if (bullishSignals > bearishSignals) {
-    return { status: "Moderately Bullish", color: "text-green-600", borderColor: "border-l-green-500" };
-  } else if (bearishSignals > bullishSignals) {
-    return { status: "Moderately Bearish", color: "text-red-600", borderColor: "border-l-red-500" };
-  } else {
-    return { status: "Neutral", color: "text-yellow-600", borderColor: "border-l-yellow-500" };
-  }
-}
-
-// ============================================================================
 // MAIN COMPONENT
 // ============================================================================
 
@@ -565,7 +112,7 @@ export default function SymbolAnalysisPage() {
   const [hasMounted, setHasMounted] = useState<boolean>(false);
 
   // State with Option types
-  const { supabase, user, isLoading: isAuthLoading } = useAuth();
+  const { user, isLoading: isAuthLoading } = useAuth();
 
   useEffect(() => {
     setHasMounted(true);
@@ -576,29 +123,20 @@ export default function SymbolAnalysisPage() {
       router.push("/");
     }
   }, [user, isAuthLoading, router, hasMounted]);
-  const [symbolValid, setSymbolValid] = useState<boolean | null>(null); // null = checking, true = valid, false = invalid
-  const [profile, setProfile] = useState<Option.Option<ProfileDBRow>>(Option.none());
-  const [quote, setQuote] = useState<Option.Option<Database["public"]["Tables"]["live_quote_indicators"]["Row"]>>(Option.none());
-  const [ratios, setRatios] = useState<Option.Option<RatiosTtmDBRow>>(Option.none());
-
-  // Insider trading data (arrays for multiple records)
-  const [insiderTransactions, setInsiderTransactions] = useState<InsiderTransactionsDBRow[]>([]);
-
-  // Valuations data (array for historical DCF data)
-  const [valuations, setValuations] = useState<ValuationsDBRow[]>([]);
-
-  // Financial statements data (latest statement for ROIC and FCF Yield calculations)
-  const [financialStatement, setFinancialStatement] = useState<Option.Option<FinancialStatementDBRow>>(Option.none());
-  // Historical financial statements for ROIC history chart
-  const [financialStatementsHistory, setFinancialStatementsHistory] = useState<FinancialStatementDBRow[]>([]);
-
-  // Global market data (for WACC calculations)
-  const [marketRiskPremiums, setMarketRiskPremiums] = useState<MarketRiskPremiumDBRow[]>([]);
-  const [treasuryRates, setTreasuryRates] = useState<TreasuryRateDBRow[]>([]);
-
-  // Analyst data (for Contrarian Indicators)
-  const [gradesHistorical, setGradesHistorical] = useState<GradesHistoricalDBRow[]>([]);
-  const [analystPriceTargets, setAnalystPriceTargets] = useState<Option.Option<AnalystPriceTargetsDBRow>>(Option.none());
+  const {
+    symbolValid,
+    profile,
+    quote,
+    ratios,
+    insiderTransactions,
+    valuations,
+    financialStatement,
+    financialStatementsHistory,
+    marketRiskPremiums,
+    treasuryRates,
+    gradesHistorical,
+    analystPriceTargets,
+  } = useSymbolAnalysisData(ticker);
 
   // Determine relevant cards based on data shown on this page
   // MUST be before any conditional returns to comply with Rules of Hooks
@@ -1021,447 +559,6 @@ export default function SymbolAnalysisPage() {
     contrarianIndicators.priceTarget,
     valuationMetrics.currentPrice
   );
-
-  // Memoize callbacks to prevent infinite re-renders
-  const handleProfileUpdate = useCallback((profileData: ProfileDBRow) => {
-    setProfile(Option.some(profileData));
-  }, []);
-
-  const handleQuoteUpdate = useCallback((quoteData: Database["public"]["Tables"]["live_quote_indicators"]["Row"]) => {
-    setQuote(Option.some(quoteData));
-  }, []);
-
-  const handleRatiosUpdate = useCallback((ratiosData: RatiosTtmDBRow) => {
-    setRatios(Option.some(ratiosData));
-  }, []);
-
-  const handleInsiderStatisticsUpdate = useCallback(() => {
-    return null
-  }, []);
-
-  const handleInsiderTransactionsUpdate = useCallback((transactionData: InsiderTransactionsDBRow) => {
-    setInsiderTransactions((prev) => {
-      const existing = prev.findIndex(
-        (t) =>
-          t.symbol === transactionData.symbol &&
-          t.filing_date === transactionData.filing_date &&
-          t.reporting_cik === transactionData.reporting_cik &&
-          t.securities_transacted === transactionData.securities_transacted
-      );
-      if (existing >= 0) {
-        const updated = [...prev];
-        updated[existing] = transactionData;
-        return updated.sort((a, b) => {
-          const aDate = a.transaction_date || a.filing_date;
-          const bDate = b.transaction_date || b.filing_date;
-          return new Date(bDate).getTime() - new Date(aDate).getTime();
-        });
-      }
-      return [...prev, transactionData].sort((a, b) => {
-        const aDate = a.transaction_date || a.filing_date;
-        const bDate = b.transaction_date || b.filing_date;
-        return new Date(bDate).getTime() - new Date(aDate).getTime();
-      });
-    });
-  }, []);
-
-  const handleValuationsUpdate = useCallback((valuationData: ValuationsDBRow) => {
-    setValuations((prev) => {
-      const existing = prev.findIndex(
-        (v) =>
-          v.symbol === valuationData.symbol &&
-          v.date === valuationData.date &&
-          v.valuation_type === valuationData.valuation_type
-      );
-      if (existing >= 0) {
-        const updated = [...prev];
-        updated[existing] = valuationData;
-        return updated.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-      }
-      return [...prev, valuationData].sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-    });
-  }, []);
-
-  const handleFinancialStatementUpdate = useCallback((statementData: FinancialStatementDBRow) => {
-    // Only update if this is the latest statement (by date)
-    setFinancialStatement((prev) => {
-      if (Option.isNone(prev)) {
-        return Option.some(statementData);
-      }
-      const prevDate = new Date(prev.value.date);
-      const newDate = new Date(statementData.date);
-      // Update if new statement is more recent
-      if (newDate >= prevDate) {
-        return Option.some(statementData);
-      }
-      return prev;
-    });
-  }, []);
-
-  const handleGradesHistoricalUpdate = useCallback((gradesData: GradesHistoricalDBRow) => {
-    setGradesHistorical((prev) => {
-      const existing = prev.findIndex(
-        (g) => g.symbol === gradesData.symbol && g.date === gradesData.date
-      );
-      if (existing >= 0) {
-        const updated = [...prev];
-        updated[existing] = gradesData;
-        return updated.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-      }
-      return [...prev, gradesData].sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-    });
-  }, []);
-
-  const handleAnalystPriceTargetsUpdate = useCallback((payload: { new: AnalystPriceTargetsDBRow | null; old: AnalystPriceTargetsDBRow | null }) => {
-    if (payload.new) {
-      setAnalystPriceTargets(Option.some(payload.new));
-    } else if (payload.old && !payload.new) {
-      setAnalystPriceTargets(Option.none());
-    }
-  }, []);
-
-  // Use the existing useStockData hook for Realtime subscriptions
-  // MUST be called before any conditional returns to follow Rules of Hooks
-  useStockData({
-    symbol: symbolValid === true ? ticker : "", // Only subscribe if symbol is valid
-    onProfileUpdate: handleProfileUpdate,
-    onLiveQuoteUpdate: handleQuoteUpdate,
-    onRatiosTTMUpdate: handleRatiosUpdate,
-    onInsiderTradingStatisticsUpdate: handleInsiderStatisticsUpdate,
-    onInsiderTransactionsUpdate: handleInsiderTransactionsUpdate,
-    onValuationsUpdate: handleValuationsUpdate,
-    onFinancialStatementUpdate: handleFinancialStatementUpdate,
-    onGradesHistoricalUpdate: handleGradesHistoricalUpdate,
-  });
-
-  // Subscribe to analyst_price_targets separately (not in useStockData yet)
-  // MUST be called before any conditional returns to follow Rules of Hooks
-  useEffect(() => {
-    if (!supabase || !ticker || symbolValid !== true) return;
-
-    const channel = supabase
-      .channel(`analyst-price-targets-${ticker}`)
-      .on<AnalystPriceTargetsDBRow>(
-        "postgres_changes",
-        {
-          event: "*",
-          schema: "public",
-          table: "analyst_price_targets",
-          filter: `symbol=eq.${ticker}`,
-        },
-        (payload) => {
-          handleAnalystPriceTargetsUpdate({
-            new: (payload.new as AnalystPriceTargetsDBRow) || null,
-            old: (payload.old as AnalystPriceTargetsDBRow) || null,
-          });
-        }
-      )
-      .subscribe();
-
-    // Fetch initial data
-    supabase
-      .from("analyst_price_targets")
-      .select("*")
-      .eq("symbol", ticker)
-      .maybeSingle()
-      .then(({ data, error }) => {
-        if (error) {
-          console.error(`[SymbolAnalysisPage] Error fetching analyst price targets:`, error);
-          return;
-        }
-        if (data) {
-          handleAnalystPriceTargetsUpdate({ new: data, old: null });
-        }
-      });
-
-    // Fetch initial grades_historical
-    supabase
-      .from("grades_historical")
-      .select("*")
-      .eq("symbol", ticker)
-      .order("date", { ascending: false })
-      .limit(10)
-      .then(({ data, error }) => {
-        if (error) {
-          console.error(`[SymbolAnalysisPage] Error fetching grades historical:`, error);
-          return;
-        }
-        if (data) {
-          setGradesHistorical(data);
-        }
-      });
-
-    return () => {
-      supabase.removeChannel(channel);
-    };
-  }, [supabase, ticker, symbolValid, handleAnalystPriceTargetsUpdate]);
-
-  // Subscribe to global tables (market_risk_premiums, treasury_rates) for WACC
-  // MUST be called before any conditional returns to follow Rules of Hooks
-  useEffect(() => {
-    if (!supabase) return;
-
-    // Subscribe to market_risk_premiums (global table, no symbol filter)
-    const mrpChannel = supabase
-      .channel("market-risk-premiums-global")
-      .on<MarketRiskPremiumDBRow>(
-        "postgres_changes",
-        {
-          event: "*",
-          schema: "public",
-          table: "market_risk_premiums",
-        },
-        (payload: MarketRiskPremiumPayload) => {
-          if (payload.new) {
-            const newRecord = payload.new as MarketRiskPremiumDBRow;
-            setMarketRiskPremiums((prev): MarketRiskPremiumDBRow[] => {
-              const existing = prev.findIndex((m) => m.country === newRecord.country);
-              if (existing >= 0) {
-                const updated = [...prev];
-                updated[existing] = newRecord;
-                return updated;
-              }
-              return [...prev, newRecord];
-            });
-          } else if (payload.old && !payload.new) {
-            const oldRecord = payload.old as MarketRiskPremiumDBRow;
-            setMarketRiskPremiums((prev): MarketRiskPremiumDBRow[] =>
-              prev.filter((m) => m.country !== oldRecord.country)
-            );
-          }
-        }
-      )
-      .subscribe();
-
-    // Subscribe to treasury_rates (global table, no symbol filter)
-    const trChannel = supabase
-      .channel("treasury-rates-global")
-      .on<TreasuryRateDBRow>(
-        "postgres_changes",
-        {
-          event: "*",
-          schema: "public",
-          table: "treasury_rates",
-        },
-        (payload) => {
-          if (payload.new) {
-            const newRecord = payload.new as TreasuryRateDBRow;
-            setTreasuryRates((prev): TreasuryRateDBRow[] => {
-              const existing = prev.findIndex((t) => t.date === newRecord.date);
-              if (existing >= 0) {
-                const updated = [...prev];
-                updated[existing] = newRecord;
-                return updated.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-              }
-              return [...prev, newRecord].sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-            });
-          } else if (payload.old && !payload.new) {
-            const oldRecord = payload.old as TreasuryRateDBRow;
-            setTreasuryRates((prev): TreasuryRateDBRow[] =>
-              prev.filter((t) => t.date !== oldRecord.date)
-            );
-          }
-        }
-      )
-      .subscribe();
-
-    // Fetch initial data for global tables
-    Promise.all([
-      supabase
-        .from("market_risk_premiums")
-        .select("*")
-        .then(({ data, error }) => {
-          if (error) {
-            console.error(`[SymbolAnalysisPage] Error fetching market risk premiums:`, error);
-            return;
-          }
-          if (data) {
-            setMarketRiskPremiums(data);
-          }
-        }),
-      supabase
-        .from("treasury_rates")
-        .select("*")
-        .order("date", { ascending: false })
-        .limit(30) // Get last 30 days
-        .then(({ data, error }) => {
-          if (error) {
-            console.error(`[SymbolAnalysisPage] Error fetching treasury rates:`, error);
-            return;
-          }
-          if (data) {
-            setTreasuryRates(data);
-          }
-        }),
-    ]);
-
-    return () => {
-      supabase.removeChannel(mrpChannel);
-      supabase.removeChannel(trChannel);
-    };
-  }, [supabase]);
-
-  // Fetch historical financial statements for ROIC chart
-  // MUST be called before any conditional returns to follow Rules of Hooks
-  useEffect(() => {
-    if (!supabase || !ticker || symbolValid !== true) return;
-
-    supabase
-      .from("financial_statements")
-      .select("*")
-      .eq("symbol", ticker)
-      .eq("period", "FY") // Only annual reports
-      .order("fetched_at", { ascending: false })
-      .order("date", { ascending: false })
-      .limit(20) // Get more than needed, then deduplicate
-      .then(({ data, error }) => {
-        if (error) {
-          console.error(`[SymbolAnalysisPage] Error fetching financial statements history:`, error);
-          return;
-        }
-        if (data) {
-          // Deduplicate: keep only the latest entry (by fetched_at) for each fiscal_year
-          const deduplicated = data
-            .reduce((acc: FinancialStatementDBRow[], item: FinancialStatementDBRow) => {
-              const existing = acc.find((a) => a.fiscal_year === item.fiscal_year);
-              if (!existing) {
-                acc.push(item);
-              } else {
-                // Keep the one with the latest fetched_at
-                const existingFetchedAt = new Date(existing.fetched_at).getTime();
-                const itemFetchedAt = new Date(item.fetched_at).getTime();
-                if (itemFetchedAt > existingFetchedAt) {
-                  const index = acc.indexOf(existing);
-                  acc[index] = item;
-                }
-              }
-              return acc;
-            }, [])
-            .sort((a: FinancialStatementDBRow, b: FinancialStatementDBRow) => {
-              // Sort by date descending
-              return new Date(b.date).getTime() - new Date(a.date).getTime();
-            })
-            .slice(0, 12); // Take latest 12 quarters (3 years)
-
-          setFinancialStatementsHistory(deduplicated);
-        }
-      });
-  }, [supabase, ticker, symbolValid]);
-
-  // Fetch initial insider trading data, valuations, and financials
-  // MUST be called before any conditional returns to follow Rules of Hooks
-  useEffect(() => {
-    if (!supabase || !ticker || symbolValid !== true) return;
-
-    // 1. Fetch insider trading statistics
-
-    // 2. Fetch insider transactions (Last 6 Months, No Limit)
-    const sixMonthsAgo = new Date();
-    sixMonthsAgo.setMonth(sixMonthsAgo.getMonth() - 6);
-    const sixMonthsAgoStr = sixMonthsAgo.toISOString().split('T')[0];
-
-    supabase
-      .from("insider_transactions")
-      .select("*")
-      .eq("symbol", ticker)
-      .gte("transaction_date", sixMonthsAgoStr)
-      .order("transaction_date", { ascending: false, nullsFirst: false })
-      .order("filing_date", { ascending: false })
-      .then(({ data, error }) => {
-        if (error) console.error(`[SymbolAnalysisPage] Error fetching insider transactions:`, error);
-        if (data) setInsiderTransactions(data);
-      });
-
-    // 3. Fetch valuations (DCF - last 180 days for chart)
-    supabase
-      .from("valuations")
-      .select("*")
-      .eq("symbol", ticker)
-      .eq("valuation_type", "dcf")
-      .order("date", { ascending: false })
-      .limit(180)
-      .then(({ data, error }) => {
-        if (error) console.error(`[SymbolAnalysisPage] Error fetching valuations:`, error);
-        if (data) setValuations(data);
-      });
-
-    // 4. Fetch latest financial statement for Safety metrics
-    supabase
-      .from("financial_statements")
-      .select("*")
-      .eq("symbol", ticker)
-      .order("date", { ascending: false })
-      .order("fetched_at", { ascending: false })
-      .limit(1)
-      .then(({ data, error }) => {
-        if (error) console.error(`[SymbolAnalysisPage] Error fetching latest financial statement:`, error);
-        if (data && data.length > 0) setFinancialStatement(Option.some(data[0]));
-      });
-
-    // 5. Fetch historical financial statements for ROIC history chart
-    supabase
-      .from("financial_statements")
-      .select("*")
-      .eq("symbol", ticker)
-      .eq("period", "FY") // Only annual reports
-      .order("fetched_at", { ascending: false })
-      .order("date", { ascending: false })
-      .then(({ data, error }) => {
-        if (error) {
-          console.error(`[SymbolAnalysisPage] Error fetching financial statements history:`, error);
-          return;
-        }
-        if (data) {
-          const deduplicated = data.reduce((acc: typeof data, current: typeof data[0]) => {
-            const existing = acc.find((item: typeof data[0]) => item.fiscal_year === current.fiscal_year);
-            if (!existing) {
-              acc.push(current);
-            } else {
-              const existingFetchedAt = existing.fetched_at ? new Date(existing.fetched_at).getTime() : 0;
-              const currentFetchedAt = current.fetched_at ? new Date(current.fetched_at).getTime() : 0;
-              if (currentFetchedAt > existingFetchedAt) {
-                const index = acc.indexOf(existing);
-                acc[index] = current;
-              }
-            }
-            return acc;
-          }, [] as typeof data);
-
-          const sorted = deduplicated
-            .sort((a: typeof data[0], b: typeof data[0]) => new Date(b.date).getTime() - new Date(a.date).getTime())
-            .slice(0, 12);
-
-          setFinancialStatementsHistory(sorted);
-        }
-      });
-  }, [supabase, ticker, symbolValid]);
-
-  // Validate symbol exists in listed_symbols before loading
-  // MUST be called before any conditional returns to follow Rules of Hooks
-  useEffect(() => {
-    if (!supabase || !ticker) return;
-
-    // Check if symbol exists in listed_symbols
-    supabase
-      .from("listed_symbols")
-      .select("symbol")
-      .eq("symbol", ticker)
-      .maybeSingle()
-      .then(({ data, error }) => {
-        if (error) {
-          console.error(`[SymbolAnalysisPage] Error checking symbol validity:`, error);
-          setSymbolValid(true); // Allow page to load on error (graceful degradation)
-          return;
-        }
-
-        if (!data) {
-          // Symbol not found
-          setSymbolValid(false);
-        } else {
-          setSymbolValid(true);
-        }
-      });
-  }, [supabase, ticker]);
 
   // Don't render page content until symbol is validated
   if (symbolValid === null) {
@@ -2329,45 +1426,3 @@ export default function SymbolAnalysisPage() {
 
 // ============================================================================
 // MICRO COMPONENTS
-// ============================================================================
-
-interface ScorecardItemProps {
-  icon: React.ReactNode;
-  label: string;
-  status: string;
-  statusColor: string;
-}
-
-function ScorecardItem({ icon, label, status, statusColor }: ScorecardItemProps) {
-  return (
-    <div className="flex flex-col justify-center">
-      <div className="flex items-center gap-2 text-muted-foreground mb-1 text-xs font-medium uppercase tracking-wider">
-        {icon} {label}
-      </div>
-      <div className={cn("font-bold text-lg truncate", statusColor)}>
-        {status}
-      </div>
-    </div>
-  );
-}
-
-interface MetricRowProps {
-  label: string;
-  value: string | null;
-  subtext?: string;
-  highlight?: boolean;
-}
-
-function MetricRow({ label, value, subtext, highlight }: MetricRowProps) {
-  return (
-    <div className="flex justify-between items-center py-1">
-      <div>
-        <div className="text-sm text-muted-foreground">{label}</div>
-        {subtext && <div className="text-xs text-muted-foreground/70">{subtext}</div>}
-      </div>
-      <div className={cn("text-base font-medium", highlight && "text-xl font-bold text-primary")}>
-        {value || <div className="h-4 w-16 bg-muted animate-pulse rounded" />}
-      </div>
-    </div>
-  );
-}
